cpu/idt: separate stage2 and SVSM IDT
Stage2 and SVSM kernel have to handle interrupts in different ways.

This creates 2 different IDT handler arrays: one calling stage2
handlers, and one calling SVSM handlers.

cpu/idt/stage2.rs and cpu/idt/svsm.rs expose the same functions,
callable from the appropriate context (stage2 or SVSM).
