+++ src/cpu/idt/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+//
+// Copyright (c) 2022-2023 SUSE LLC
+//
+// Author: Thomas Leroy <tleroy@suse.de>
+
+pub mod common;
+pub mod stage2;
+pub mod svsm;
