+++ src/debug/stacktrace.rs
@@ -6,8 +6,8 @@
 
 use crate::address::{Address, VirtAddr};
 #[cfg(feature = "enable-stacktrace")]
-use crate::cpu::idt::is_exception_handler_return_site;
-use crate::cpu::idt::X86ExceptionContext;
+use crate::cpu::idt::common::is_exception_handler_return_site;
+use crate::cpu::idt::common::X86ExceptionContext;
 use crate::cpu::percpu::this_cpu;
 #[cfg(feature = "enable-stacktrace")]
 use crate::mm::address_space::STACK_SIZE;
