+++ src/stage2.rs
@@ -15,6 +15,8 @@ use core::slice;
 use svsm::address::{Address, PhysAddr, VirtAddr};
 use svsm::console::{init_console, install_console_logger, WRITER};
 use svsm::cpu::cpuid::{dump_cpuid_table, register_cpuid_table, SnpCpuidTable};
+use svsm::cpu::gdt::load_gdt;
+use svsm::cpu::idt::stage2::early_idt_init;
 use svsm::cpu::percpu::{this_cpu_mut, PerCpu};
 use svsm::elf;
 use svsm::fw_cfg::FwCfg;
@@ -81,6 +83,9 @@ static CONSOLE_SERIAL: SerialPort = SerialPort {
 };
 
 fn setup_env() {
+    load_gdt();
+    early_idt_init();
+
     install_console_logger("Stage2");
     init_kernel_mapping_info(
         VirtAddr::null(),
